% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi_calc.R
\name{bmi_calc}
\alias{bmi_calc}
\title{BMI calculator}
\usage{
bmi_calc(data)
}
\arguments{
\item{data}{A data frame with height and weight}
}
\value{
data frame with height, weight, and bmi.
}
\description{
BMI calculator
}
